Split in two separate plugin (to handle request / response at different time)
